// src/utils/Utils.d.ts
export declare class Point {
    x: number;
    y: number;
    constructor(x?: number, y?: number);
    toString(): string;
}
// ...
export declare class TextureHelper {
    static createGradientCanvas: (size: PIXI.Point, colors: Array<string>, points: Array<PIXI.Point>, shadow: Shadow) => HTMLCanvasElement;
    static createRoundedGradientCanvas: (size: PIXI.Point, colors: Array<string>, points: Array<PIXI.Point>, radius?: number, shadow?: Shadow) => HTMLCanvasElement;
    static createRoundedCanvas: (width: number, height: number, fill: string, strokeColor: string, strokeWidth: number, radius: number, offset?: number) => HTMLCanvasElement;
    static createRoundedTexture: (width: number, height: number, fill: string, strokeColor: string, strokeWidth: number, radius: number, offset?: number) => Promise<PIXI.Texture>;
    static createRoundedTexture2: (width: number, height: number, fill: string, strokeColor: string, strokeWidth: number, radius: number, offset?: number) => PIXI.Texture;
    static createFillCanvas: (size: PIXI.Point, color: string, shadow?: Shadow) => HTMLCanvasElement;
    static createFillCanvasEx: (size: PIXI.Point, color: string, shadow?: Shadow) => HTMLCanvasElement;
}
// ...
export declare class Shadow extends Object {
    shadowColor: string;
    shadowBlur: number;
    shadowOffsetX: number;
    shadowOffsetY: number;
    constructor();
}